refactor: jsファイルのディレクトリを整理
