+++ app/javascript/javascripts/diagnosis/like_modal.js
