+++ app/javascript/javascripts/shared/description.js
@@ -1,13 +1,3 @@
-document.addEventListener('turbolinks:load', () => {
-  const button = document.querySelector('#button');
-
-  button.addEventListener('click', event => {
-    bars.classList.toggle('hidden');
-    xmark.classList.toggle('hidden');
-    menu.classList.toggle('translate-x-full');
-  })
-})
-
 document.addEventListener("turbolinks:load", () => {
   const toggleBtns = document.querySelectorAll('.toggle-btn');
   const descriptions = document.querySelectorAll('.description');
