+++ app/javascript/javascripts/shared/hamburger_menu.js
@@ -1,0 +1,9 @@
+document.addEventListener('turbolinks:load', () => {
+  const button = document.querySelector('#button');
+
+  button.addEventListener('click', event => {
+    bars.classList.toggle('hidden');
+    xmark.classList.toggle('hidden');
+    menu.classList.toggle('translate-x-full');
+  })
+})
